Silence warnings

+++ src/core/vscore.cpp
@@ -1191,8 +1191,8 @@ int VSNode::setLinear() {
     cacheLinear = true;
     cacheOverride = true;
     cacheEnabled = true;
-    cache.setFixedSize(true);   
-    cache.setMaxFrames(core->threadPool->threadCount() * 2 + 20);
+    cache.setFixedSize(true);
+    cache.setMaxFrames(static_cast<int>(core->threadPool->threadCount()) * 2 + 20);
     registerCache(cacheEnabled);
     return cache.getMaxFrames() / 2;
 }
